feat: v1.2.0 Major API improvements
* Plot-System plugin is no longer dependent to this plugin,
  instead a connector plugin (e.g. Plot-System) is required
  to register its data provider for managing plots.

* Fully documented the API module

* Rewrote README.md

* Renamed the command `/plot` to `/plotctl` as these commands
  is only intended for internal usage only, a new public command
  named /plot is planned to be published

* New `/review` command for editing/sending review message
  and archiving the reviewed plots

* New PlotNotificationEvent for all plot-related notification events
  using NotificationType enum.

* New PlotCreateData and PlotCreateProvider as a provider for plot data
  because we just made Plot-System out of dependent.

* Implemented to the new PlotCreateData replacing PlotEntry that
  was retrieving data from Plot-System database

* Deleted coordinates conversion utilities as it is
  now provided by the API's provider

* Deleted many unused class in the database package

* Review can now be edited with markdown and attachment(s)
  using the new /review command

* Review message will now be sending with ComponentV2

* Plot archiving now implemented directly to the review command

* Help button now sending via ComponentV2 thanks the new ReviewComponent

* Plot forum now add and remove member according to who reclaim and
  claimed at the time

+++ discord-plotsystem/src/main/java/asia/buildtheearth/asean/discord/plotsystem/commands/AbstractPlotArchiveCommand.java
@@ -104,7 +104,8 @@ abstract sealed class AbstractPlotArchiveCommand
 
     protected Consumer<File> getCacheDeleter(MessageChannel channel) {
         return file -> {
-            if (file.delete()) channel.sendMessage(
+            if(!file.getName().startsWith(Constants.PLOT_IMAGE_FILE)) return;
+            if(file.delete()) channel.sendMessage(
                 getLang(MESSAGE_MEDIA_CACHE_DELETED)
                 .replace(Format.FILENAME, file.getName()))
                 .queue();
